requirements.txt
fastapi>=0.68.0
uvicorn
sqlalchemy
pydantic>=2.0.0
pydantic-settings
python-dotenv
stripe
python-jose[cryptography]
boto3
elasticsearch
pika
redis
auth0-python==4.7.2
cryptography>=43.0.1,<44.0.0
python-multipart
